Move save_content_to_csv to utils module
